Adding support for GTK3

+++ src/common/Lgi/GFileSelect.cpp
@@ -1178,7 +1178,7 @@ int GFileSelectDlg::OnNotify(GViewI *Ctrl, int Flags)
 							Ext++;
 							char s[256];
 							ZeroObj(s);
-							memcpy(s, File, (int)Ext-(int)File);
+							memcpy(s, File, Ext-File);
 							strcat(s, DefExt);
 
 							OnFile(s);
